requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
APScheduler
plyer
pystray
pillow
requests
python-dotenv
rich
reportlab
weasyprint
Jinja2
sqlite-utils
pyyaml
scikit-learn
xgboost
matplotlib
seaborn
joblib